cut down agent cruft
