+++ src/routes/(main)/activities/berkshires/+page.svelte
@@ -4,7 +4,7 @@
 	import HR from '$lib/components/HR.svelte';
 	import PlaceMap from '$lib/components/PlaceMap.svelte';
 	import InteractivePlaceMap from '$lib/components/InteractivePlaceMap.svelte';
-	import { getPlacesFromItinerary, getPlaceUrl } from '$lib/places.js';
+	import { getPlacesFromItinerary, getPlaceUrl, groupPlaces } from '$lib/places.js';
 
 	let { data } = $props();
 	let { itineraries, places } = data;
@@ -12,24 +12,7 @@
 	const bodyLength = (body) =>
 		body?.flatMap((block) => block.children?.map((span) => span.text) ?? []).join('').length ?? 0;
 
-	// Group places by their tags. Places with multiple tags appear in each group.
-	// Places with no tags go into an "Other" bucket.
-	const placesByTag = $derived(() => {
-		const groups = {};
-		for (const place of places) {
-			const tags = place.tags?.length ? place.tags : ['Other'];
-			for (const tag of tags) {
-				if (!groups[tag]) groups[tag] = [];
-				groups[tag].push(place);
-			}
-		}
-		// Sort tags alphabetically, but keep "Other" last
-		return Object.entries(groups).sort(([a], [b]) => {
-			if (a === 'Other') return 1;
-			if (b === 'Other') return -1;
-			return a.localeCompare(b);
-		});
-	});
+	const groups = $derived(groupPlaces(places));
 </script>
 
 <svelte:head>
@@ -86,31 +69,19 @@
 
 	<InteractivePlaceMap {places} />
 
-	<div class="place-groups">
-		{#each placesByTag() as [tag, tagPlaces]}
-			<div class="place-group">
-				<h4 class="tag-heading">{tag}</h4>
-				<ul class="place-list">
-					{#each tagPlaces as place}
-						<li class="place-item">
-							<a
-								class="place-name"
-								href={getPlaceUrl(place)}
-								target="_blank"
-								rel="noopener noreferrer"
-							>
-								{place.name}
-							</a>
-							{#if place.address?.city}
-								<span class="place-city">{place.address.city}</span>
-							{/if}
-							{#if place.description}
-								<span class="place-desc">{place.description}</span>
-							{/if}
-						</li>
-					{/each}
-				</ul>
-			</div>
+	<div class="index">
+		{#each groups as [tag, places]}
+			<h4 class="tag-heading">{tag}</h4>
+			{#each places as place}
+				<div class="item">
+					<a href={getPlaceUrl(place)} target="_blank" rel="noopener noreferrer">
+						{place.name}
+					</a>
+					{#if place.description}
+						<span class="place-desc">{place.description}</span>
+					{/if}
+				</div>
+			{/each}
 		{/each}
 	</div>
 
@@ -149,63 +120,6 @@
 	}
 	:global(.other p) {
 		margin: 0;
-	}
-
-	h2 {
-		margin-top: 2em;
-		margin-bottom: 0.5em;
-	}
-
-	.place-groups {
-		display: grid;
-		grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-		gap: 1.5em 2em;
-		margin-top: 1.5em;
-	}
-
-	.tag-heading {
-		margin: 0 0 0.4em 0;
-		font-size: 0.8rem;
-		text-transform: uppercase;
-		letter-spacing: 0.06em;
-		color: #888;
-	}
-
-	.place-list {
-		list-style: none;
-		margin: 0;
-		padding: 0;
-		display: flex;
-		flex-direction: column;
-		gap: 0.6em;
-	}
-
-	.place-item {
-		display: flex;
-		flex-direction: column;
-		gap: 1px;
-	}
-
-	.place-name {
-		font-size: 0.9rem;
-		font-weight: 600;
-		color: inherit;
-		text-decoration: none;
-	}
-	.place-name:hover {
-		color: var(--blue);
-		text-decoration: underline;
-	}
-
-	.place-city {
-		font-size: 0.78rem;
-		color: #888;
-	}
-
-	.place-desc {
-		font-size: 0.78rem;
-		color: #555;
-		font-style: italic;
 	}
 
 	footer {
